replaces functionality of isLogicalCombination with isContextFunctionalPseudoClass
note: possible logic error in selector-max-type, will investigate further
