+++ lib/reference/keywordSets.js
@@ -300,6 +300,8 @@ keywordSets.webkitProprietaryPseudoClasses = new Set([
 keywordSets.pseudoClasses = uniteSets(
 	keywordSets.aNPlusBNotationPseudoClasses,
 	keywordSets.linguisticPseudoClasses,
+	keywordSets.logicalCombinationsPseudoClasses,
+	keywordSets.aNPlusBOfSNotationPseudoClasses,
 	keywordSets.otherPseudoClasses,
 );
