+++ lib/rules/selector-max-attribute/index.js
@@ -3,7 +3,7 @@
 'use strict';
 
 const _ = require('lodash');
-const isLogicalCombination = require('../../utils/isLogicalCombination');
+const isContextFunctionalPseudoClass = require('../../utils/isContextFunctionalPseudoClass');
 const isStandardSyntaxRule = require('../../utils/isStandardSyntaxRule');
 const optionsMatches = require('../../utils/optionsMatches');
 const parseSelector = require('../../utils/parseSelector');
@@ -49,8 +49,8 @@ function rule(max, options) {
 
 		function checkSelector(selectorNode, ruleNode) {
 			const count = selectorNode.reduce((total, childNode) => {
-				// Only traverse inside actual selectors and logical combinations
-				if (childNode.type === 'selector' || isLogicalCombination(childNode)) {
+				// Only traverse inside actual selectors  and context functional pseudo classes
+				if (childNode.type === 'selector' || isContextFunctionalPseudoClass(childNode)) {
 					checkSelector(childNode, ruleNode);
 				}
