+++ lib/rules/selector-max-compound-selectors/index.js
@@ -2,7 +2,7 @@
 
 'use strict';
 
-const isLogicalCombination = require('../../utils/isLogicalCombination');
+const isContextFunctionalPseudoClass = require('../../utils/isContextFunctionalPseudoClass');
 const isStandardSyntaxRule = require('../../utils/isStandardSyntaxRule');
 const parseSelector = require('../../utils/parseSelector');
 const report = require('../../utils/report');
@@ -39,8 +39,8 @@ function rule(max) {
 			let compoundCount = 1;
 
 			selectorNode.each((childNode) => {
-				// Only traverse inside actual selectors and logical combinations
-				if (childNode.type === 'selector' || isLogicalCombination(childNode)) {
+				// Only traverse inside actual selectors  and context functional pseudo classes
+				if (childNode.type === 'selector' || isContextFunctionalPseudoClass(childNode)) {
 					checkSelector(childNode, rule);
 				}
