+++ lib/rules/selector-max-id/index.js
@@ -3,7 +3,7 @@
 'use strict';
 
 const _ = require('lodash');
-const isLogicalCombination = require('../../utils/isLogicalCombination');
+const isContextFunctionalPseudoClass = require('../../utils/isContextFunctionalPseudoClass');
 const isStandardSyntaxRule = require('../../utils/isStandardSyntaxRule');
 const optionsMatches = require('../../utils/optionsMatches');
 const parseSelector = require('../../utils/parseSelector');
@@ -47,10 +47,10 @@ function rule(max, options) {
 
 		function checkSelector(selectorNode, ruleNode) {
 			const count = selectorNode.reduce((total, childNode) => {
-				// Only traverse inside actual selectors and logical combinations that are not part of ignored functional pseudo-classes
+				// Only traverse inside actual selectors  and context functional pseudo classes that are not part of ignored functional pseudo-classes
 				if (
 					childNode.type === 'selector' ||
-					(isLogicalCombination(childNode) &&
+					(isContextFunctionalPseudoClass(childNode) &&
 						!isIgnoredContextFunctionalPseudoClass(childNode, options))
 				) {
 					checkSelector(childNode, ruleNode);
