+++ lib/rules/selector-max-pseudo-class/index.js
@@ -2,7 +2,7 @@
 
 'use strict';
 
-const isLogicalCombination = require('../../utils/isLogicalCombination');
+const isContextFunctionalPseudoClass = require('../../utils/isContextFunctionalPseudoClass');
 const isStandardSyntaxRule = require('../../utils/isStandardSyntaxRule');
 const keywordSets = require('../../reference/keywordSets');
 const parseSelector = require('../../utils/parseSelector');
@@ -35,8 +35,8 @@ function rule(max) {
 
 		function checkSelector(selectorNode, ruleNode) {
 			const count = selectorNode.reduce((total, childNode) => {
-				// Only traverse inside actual selectors and logical combinations
-				if (childNode.type === 'selector' || isLogicalCombination(childNode)) {
+				// Only traverse inside actual selectors and context functional pseudo classes
+				if (childNode.type === 'selector' || isContextFunctionalPseudoClass(childNode)) {
 					checkSelector(childNode, ruleNode);
 				}
