+++ lib/rules/selector-max-type/index.js
@@ -3,8 +3,8 @@
 'use strict';
 
 const _ = require('lodash');
+const isContextFunctionalPseudoClass = require('../../utils/isContextFunctionalPseudoClass');
 const isKeyframeSelector = require('../../utils/isKeyframeSelector');
-const isLogicalCombination = require('../../utils/isLogicalCombination');
 const isOnlyWhitespace = require('../../utils/isOnlyWhitespace');
 const isStandardSyntaxRule = require('../../utils/isStandardSyntaxRule');
 const isStandardSyntaxSelector = require('../../utils/isStandardSyntaxSelector');
@@ -56,8 +56,8 @@ function rule(max, options) {
 
 		function checkSelector(selectorNode, ruleNode) {
 			const count = selectorNode.reduce((total, childNode) => {
-				// Only traverse inside actual selectors and logical combinations
-				if (childNode.type === 'selector' || isLogicalCombination(childNode)) {
+				// Only traverse inside actual selectors  and context functional pseudo classes
+				if (childNode.type === 'selector' || isContextFunctionalPseudoClass(childNode)) {
 					checkSelector(childNode, ruleNode);
 				}
