+++ lib/utils/isStandardSyntaxTypeSelector.js
@@ -27,6 +27,7 @@ module.exports = function (node) {
 		if (
 			parentType === 'pseudo' &&
 			(keywordSets.aNPlusBNotationPseudoClasses.has(normalisedParentName) ||
+				keywordSets.aNPlusBOfSNotationPseudoClasses.has(normalisedParentName) ||
 				keywordSets.linguisticPseudoClasses.has(normalisedParentName) ||
 				keywordSets.shadowTreePseudoElements.has(normalisedParentName))
 		) {
